Convert main script to python 3
Stop using mechanize to make the post request to the website that
annotates the lyrics with furigana. Using the `requests` module
instead is easy once I learnt that I needed to pass the data as a
`files` argument instead of `data` argument. The reason is that the
form is not url encoded, but it's actually a multipart form.

Here's a stackoverflow question with more detailed information:
https://stackoverflow.com/questions/52109055/utf-8-text-from-website-is
-decoded-improperly-when-using-python-3-and-requests?noredirect=1#commen
t91203208_52109055