Clean up the main script
Now that the main script doesn't use `mechanize` and is interpreted as
a python 3 script, the HTML parser now makes use of Enums. Also, the
parser logic is simpler now, thanks in large part to additional
preprocessing of the lyrics that get rid of many edge cases.

This commit deliberately doesn't add any new functionality.