Refine shortcut
Decided on a definite name, for now.

Fixed bug in which the substring '.htm' would be replaced with '.jis'
in any chosen URL, instead of only those belonging to the
animelyrics.com domain.

Changed how the .tex file is shared. It can be shared as any file can,
so it can be "Saved to Files" as before, or copied to FileBrowser
directly.